Add comparison condition for Contact information
Basically, the mid-dialog request should use the Contact header that was
sent in the SIP request when forming the Request-URI. But, the test
script of Docomo didn't use the previously received Contact header.
So, SIP engine will check all contact information when validating the
Request-URI of all the incoming requests even if it's a mid-dialog
request.

Bug: 238571302
Test: manual (verify ims-reg on target device)
Change-Id: I988aee44d06e3c619cfd971ebc633a66583a2bbf