Add a new config for local session timer
Generally, if the remote UA doesn't specify the support of the session
timer or the incoming message doesn't contain the Session-Expires
header, then UA considers this session to be permanent.
However, some carriers want to use the session timer even though the
remote device doesn't specify the support of session timer. For this
purpose, new configuration item is defined and used to identify whether
the local session timer should be supported or not.
If this configuration is set to false and there is no "timer" option tag
in the Supported or Require header and no Session-Expires header, the
successful response doesn't contain the "timer" option tag and
Session-Expires header. If it's set to true, the successful response
contains the "timer" option tag and Session-Expires header regardless of
the state of the remote UA.

Bug: 239670618
Test: atest ImsStackNativeConfigTests
Test: manual (verify ims-reg/call on target device)
Change-Id: I416337ad0619823b12e94abf0787fb3bf927ab91