(Mtc) support carrier specific header in BYE
VZW and KR carriers require their specific SIP header in BYE message
based on the termination reason. Adding SIP header in BYE message is not
fully covered by enabler side and it should be set by engine side in
some cases so engine added new apis to support this.
This commit is to use the new apis in MTC enabler.

Bug: b/235752315, b/255250733
Test: simulation verified
Test: atest passed
Change-Id: If7d27bb00edbf1a7771267c9eed0e3a3bfb90a4d