(GII-REG) Update video capability based on mobile data on/off
This is VZW & TMUS requirement.
To consider Mobile On/Off in evaluating video capability both in normal and roaming.
Enable video capability when mobile data is on.
Used already defined framework key ignore_data_enabled_changed_for_video_calls and set to false only these 2 carrier.
But this is not ending with bool, was not able to read the carrier config value. So compared explicitly.
Once this requirement is covered in framework this handling can be removed

Bug: 263439409
Test: atest ImsRegistrationTrackerTest
Change-Id: I70bd3077232f36917c739d1b3f3d096c092d5f50