(Aos) Set IMS Establishment Timer as 5 minutes
This is a CL to support AT&T requirement CDR-IMS-1210
CDR-IMS-1210 IMS Establishment Timer
VoLTE devices shall implement a timer for declaring "IMS voice not available" per 3GPP TS 24.301
section 4.3.1. This timer shall be set to 5 minutes. Device behavior when in the "IMS voice not
available" state shall follow the requirements of 3GPP TS 24.301 sections 4.5, 5.5.1.3 and 5.5.3.3.
For 5G SA, the requirements are in 3GPP TS 24.501 sections 4.3.4 and 4.9.2.

Test: Test LTE-N86-5-0702 on R&S platform
Bug: b/277709887
Bug: b/277540267
Change-Id: I272e2fcb8bd16a3e51c478323214f8f155cced6e