(Mtc) Implement : silent_redial_max_duration_millis_int config
This adds a new configuration silent_redial_max_duration_millis_int and
its implementation in SilentRedialHelper.
The timer starts when the first Silent Redial begins. If consecutive
silent redials occur and the timer expires during the retry, the next
round of silent redials will not be attempted, and the call will be
terminated as a failure.
The default value is 30s.
And, this commit also updates the default value of
silent_redial_max_retry_count_int configuration to 1.

Test: atest ImsStackNativeEnablerMtcTests
Test: simulated with various values of max duration and max retry count
Bug: 361207658
Flag: NONE pre-production
Change-Id: I562100de8016e0723fdea27198921a13981886c3