(Mtc) Set reason phrase when call is rejected because VoPS
This commit adds a configuration to set reason phrase for 488 when the
call is blocked by VoPS off

It doesn't need to be handled by MtcRoutingRejectHandler, because IMS
registration doesn't change when the VoPS on/off.

call_terminate_reason_header_vops_off_string config is changed to set
reason pharse. No requirement to set Reason header in this case.

Bug: 381989518
Test: ImsStackNativeMtcTests
Test: TISS simulation w/ test code
Flag: NONE pre-production
Change-Id: Ic9d0c8df14b04fb7616284194f6d1bac21e50f35