(config) Add a config for Expires header in REGISTER request
This change is to add a new configuration to determine whether to add
the Expires header in the REGISTER request.
Basically, the Expires header in the REGISTER request is optional, but
should be added depending on specific test case and carrier's
requirements, so add the corresponding configuration.
The default value for this configuration is set to false.

Bug: 383428523
Flag: EXEMPT bugfix
Test: atest ImsStackNativeTests
Test: manual (verified IMS-REG based on whether the Expires header is
present or not)

Change-Id: I32323f930c2db03516c4c0144984230fa1edffbe