(Mtc) Implement PRACK wait timer
This modification implements the behavior of terminating a call with
a 500 Server Internal Error if a PRACK is not received within a
certain time after sending a Reliable Provisional Response, along
with configuration options.
According to the RFC 3262 standard, if an RPR does not receive a
PRACK within T1 x 64, it should send a 500 error. Currently, the
IMS SIP engine does not support this behavior(b/232669991), so the
enabler needs to handle it. Furthermore, specific operators like
KDDI require a configurable wait time.

In this implementation, considering KDDI's requirements and user
experience, the timer will only be active for the first RPR sent

Test: verified with simulation
Test: atest ImsStackNativeEnablerMtcTests
Bug: 384410035
Flag: NONE pre-production
Change-Id: I569837081874ca34989757a015ceddfd6589706f