(Aos) Keep emergency pdn when no pcscf available
Precondition
imsemergency.keep_epdn_upon_pcscf_unavailable_bool: true
In case of requesting fake registration with the next pcscf
 - The UE needs to keep emergency pdn if no pcscf available.

The detailed scenario
This carrier that set this to true has special emergency timers in their requirement.
The timer is defined the time from INVITE to 100 response.
The UE needs to set fake registration with the next P-CSCF when receiving error response before the timer stops.
The UE needs to keep emergency pdn if there is no pcscf available in this case.

Flag: NONE pre-production
Bug: 382002932
Bug: 390504501
Test: manual test with Tiss and atest
Change-Id: I9e30035f7f660d61dcf62a70e5395fd3b6aae4fb