(AoS) Update an asset release_epdn_of_unavailable_network_bool
This commit updates assets for carriers that do not require release_epdn_of_unavailable_network_bool.
Verizon is required to maintain emergency PDN during the ECBM.
So we update the release_epdn_of_unavailable_network_bool as false to not release EPDN even if the network is unavailable.

Bug: 388576997
Test: atest ImsStackNativeEnablerAosTests
Flag: NONE pre-production
Change-Id: Ida325a461f1f90a6f2e16c430f8933a9db35f051