(Ssc) Set ut_temporary_block_timer_with_any_reason_sec_int to 1s
This is a CL that changes the temporary block timer for Ut/XCAP to 1s
when it fails due to GBA, DNS query, TCP socket, or PDN connection
problems. This is to give the UE a chance to try Ssc over Ut/XCAP as
much as possible.

Bug: 413563090
Bug: 419161840
Test: build
Flag: NONE resource only update
Change-Id: I903200fb4c54cb1cb6d7f785d1e412623893e176