(Mtc) can't display incoming call number with no PAID
Default configs are `oip_source_from_header_bool=false` and
`enable_oip_header_policy_fallback_bool=false`.
With this config, if there is only the From header and no PAID header,
this issue can happen.
The standard said like below. Based on that using the From header can be
quite risky though, use the From header for user experience if there is
no specific requirement that said not using the From header.
IR92 - 2 IMS Feature Set
2.3.12 Originating Identification Presentation
The UE must support the presentation of the originating user identity
both from the identity within the P-Asserted-Identity header field and
the identity within the From header field.
NOTE : As, by default, the identity in the From header field may not be
network asserted, it is the responsibility of the network to ensure that
the From header field contains a reliable identity of the originating
when it is sent to the UE.

Test: atest ImsStackNativeEnablerMtcTests
Test: simulating with the TISS
Bug: 417460317
Flag: NONE pre-production
Change-Id: I5c891ffd7a7246b2305f770eb148dbfab0c99374