(Mtc) Add hyphen into the IMEI value used in the PPI header
In the clause 13.13 of TS 23.003, the IMEI based identity included
in P-Preferred-Identity header shall be encoded according to
ABNF of imeival as defined in IETF RFC 7254 below:

    imeival  =  tac "-" snr "-" spare

In the DCM test case IO030212, script expects the IMEI value with
hyphen in the PPI header, but DUT currently doesn't append hyphen
into the IMEI.

To accommodate the standard and carrier operation, this commit
creates a new format "#IMEIWITHHYPHEN#" to add '-' into the IMEI
value if carrier expects to receive such string in the PPI or
PEI header.

Besides, we change the default value in carrier_config.xml
from "#IMEI#" to "#IMEIWITHHYPHEN#".

Bug: 388705348
Justification: b/388705348#comment45, b/388705348#comment55
Test: atest ImsStackNativeEnablerMtcTests
Test: Validated by carrier test case (b/388705348#comment61)
Flag: NONE pre-production
Change-Id: I966760a820e92eab7ba1e1bb38b9e6f8ddf75996