(Aos) Change default value of the release_epdn_upon_ecall_end_in_fake_mode_bool
Until now, the configuration to release E-PDN when anonymous ECall (ECall without E-Reg) ends was set to TRUE by default. However, during the GCF TC (11.2.4) validation process, it was discovered that the expected behavior is to maintain the E-PDN even after the call is terminated when an emergency call is made in No SIM state.

Therefore, we will change the default value as FALSE.
And for the following operators that require the E-PDN to be released after the emergency call is terminated, the configuration value is set to FALSE:
- JP operators, KR operators, FET TW, Dish US, TMO US, Telstra AU, Reliance IN, China-Telecom CN, Bell CA, Fredom-Mobile CA

Bug: 419445886
Test: Build and Manual test
Flag: NONE resource only update
Change-Id: I5b644e4823ddc19b15d3fbb2b543cb899c0b9029