(Mtc) add a new configuration reject_code_and_reason_code_set_string_array for emergency
This commit adds a new configuration
imsemergency.reject_code_and_reason_code_set_string_array.
This is used to determine the reason-code which is used when an
emergency call is terminated before being connected and ImsStack
notifies it to call framework.

The detail can be found in go/mtc-emergency-starterrorhandler.

Test: mm build using the same topic
Bug: 425257798
Flag: NONE pre-production
Change-Id: I87ca134f7854917f6938010da31fe2733cb6fbf7