(Aos) Correct IMPU index(0) for KDDI IMS registration
According to KDDI requirements, the device must use the 1st IMPU (index 0) from the UICC for IMS registration. The current configuration incorrectly defaults to index 1 (the 2nd IMPU).

This misconfiguration did not cause functional issues in prior tests because the 1st and 2nd IMPUs on the KDDI UICC provided the same IMSI-based identity. However, this violates the carrier specification.

This change corrects the impu_index_for_ims_registration_int value to 0 to ensure the 1st IMPU is used, aligning with KDDI's requirements.

Bug: 441176710
Test: build
Flag: NONE pre-production

Change-Id: Ia8b39aecaefddcb9c1e19061d74d125416dd150f