config: Use empty user-info for Contact header of REGISTER request
If the Contact header of REGISTER request contains a UUID-based
user-info field, the network uses this information to create a tel URI
when sending an INVITE request. This behavior causes 404 response to
incoming INVITE request.
To avoid this situation, the Contact header of the REGISTER request does
not include the user-info field, so the network uses the SIP URI in the
Contact header of the REGISTER request as the Request-URI of the SIP
request sent to UE.

Bug: 441722831
Flag: NONE resource update only
Test: manual (verify MT IMS calls)
Change-Id: I365013d453199a5fb743a271bd7569b2a024e6ae