(Mtc) Config: Remove action REDIRECTION_BY_CONTACT for 380 response in the carrier_config.xml
Previously, the device's default settings allowed for the
REDIRECTION_BY_CONTACT action, enabling it to reattempt an INVITE
request by utilizing a new address provided in the Contact header
field from the 380 response.

Per RFC 3261, Section 21.3, this retry action is not specified for
the 380 response. Furthermore, we don't find any carrier requirement
for such retry behavior as of now. We decide to exclude the action
REDIRECTION_BY_CONTACT for the 380 response in the default.xml.

Addtionally, this commit also adds new carrier_config file to set exclusive value 380:0 for Masmovil_ES device. With the customized settings, the Masmovil_ES device would directly initiate CSFB without validating the 380 response content, which is the same device behavior on the legacy Pixel phones.

Bug: 446858282
Test: Validate config change for Masmovil_ES device
Flag: NONE pre-production
Change-Id: I85b758e2bc0ae0a478dfafb86bd6345b7719ef78