(Mtc) Implement Register-Restoration on BYE Transaction Timeout
This modification implements the DTAG specific requirement (related to MTR-39418)
concerning non-INVITE client transaction timeouts.

The DTAG VoLTE 12.1 Test Case explicitly expects the UE to initiate a
Register-Restoration when a response to the BYE request is not received.

- The change adheres to 3GPP TS 24.229 5.1.2A.6, which mandates that the UE
shall not perform a CS retry upon Timer F expiry in the "Trying" state of a
non-INVITE client transaction, but instead shall initiate a Register-Restoration.
- The MTC logic is updated to leverage the `ByeTransactionHandler`
(introduced in the previous commit) to trigger the necessary Register-Restoration
action upon call termination (specifically, when the BYE transaction times out).

Test: atest ImsStackNativeEnablerMtcTests
Test: normal call termination, BYE transaction timeout,
      stop enabler by sim reloaded, forking, CANCEL.
Bug: 452167947
Flag: NONE pre-production
Change-Id: I6ffc9655a2e6fae04061ae9f06210374c73433b8