(config) Remove suffix to identify the parent carrier from the IMS asset file
IMS asset XML files contain the following suffix to identify the parent carrier.
- {original xml file name}_p{parent-carrier-id}
However, now that we are specifying the parent carrier ids in the asset separately, there is no need to include the suffix in the file name.
Therefore, the existing suffix is removed from the file name of the IMS asset.

Bug: 458564057
Test: build
Flag: NONE resource only update
Change-Id: Ia48026164318968fd7760d7c8678d40d163d8ce3