(sipcore) Ignore "udp" transport parameter for outgoing requests
The emergency call failure is observed on TW network(FET) while roaming
with TMO SIM. The carrier network sends SIP messages with the
"transport=udp" parameter in the Record-Route and Contact headers.
According to SIP routing rules, this forces large SIP requests
(mid-dialog requests) to be sent over UDP. However, the network doesn't
respond to these SIP requests, resulting in call failure.
To resolver this issue, the SIP engine has been modified to ignore the
transport parameter ("udp") even if the network unintentionally includes
it in SIP messages. This ensures that the transport protocol is
determined according to the SIM carrier's own configuration.
At this moment, the configuration will be enabled only for TMO carrier.

Bug: 460027902
Bug: 465035193
Flag: EXEMPT BUGFIX
Test: atest ImsStackNativeTests
Test: manual (verify emergency call with TMO SIM while roaming on TW
carrier's networks)

Change-Id: I06fb0abf637e3f2a45047760d6f13e09d87f23b8